refactor: moves methods to files
